Fixed bug: Python's 'pass by reference'-rule for arrays
Apparently Python's default is to pass array as reference, NOT by value.
Hence, there is NOT a copy made of the array, but the original is adjusted.

Fixed bug: After moving the breakpoints, solution X always updated to the accepted or rejected (!) trial configuration.

Fix: Use a.copy() on numpy array at the beginning of TakeStep()
NOTE: Pass starting guess, lower bound, upper etc. all as np.array() to be sure that .copy() will work
alternative: "from copy import copy" should also work on lists (not implemented, could be done if needed)